alert fix
